Extract PythonTrialQueueAdapter into its own module
