+++ rustuna_pyo3/src/study.rs
@@ -29,8 +29,9 @@ use crate::storage::PyStorage;
 use crate::trial::{PyPersistedTrial, PyTrial, PyTrialState};
 use crate::trial_queue::directory::PyDirectoryTrialQueue;
 use crate::trial_queue::inmemory::PyInMemoryTrialQueue;
+use crate::trial_queue::python::PythonTrialQueueAdapter;
 use crate::trial_queue::sqlite3::PySQLite3TrialQueue;
-use crate::trial_queue::{PyObjectTrialQueue, PyPyObjectTrialQueue};
+use crate::trial_queue::PyPyObjectTrialQueue;
 
 type SharedStorage = Arc<RwLock<dyn Storage>>;
 type SharedSampler = Arc<Mutex<dyn Sampler>>;
@@ -117,8 +118,7 @@ fn into_trial_queue_pyobj(
     match trial_queue {
         Some(trial_queue) => {
             let trial_queue_ref = trial_queue.bind(py);
-            if let Ok(py_obj_trial_queue) = trial_queue_ref.extract::<PyPyObjectTrialQueue>()
-            {
+            if let Ok(py_obj_trial_queue) = trial_queue_ref.extract::<PyPyObjectTrialQueue>() {
                 Ok((
                     py_obj_trial_queue.queue.clone() as SharedTrialQueue,
                     trial_queue.clone_ref(py),
@@ -145,7 +145,7 @@ fn into_trial_queue_pyobj(
                     trial_queue.clone_ref(py),
                 ))
             } else {
-                let queue: SharedTrialQueue = Arc::new(RwLock::new(PyObjectTrialQueue::new(
+                let queue: SharedTrialQueue = Arc::new(RwLock::new(PythonTrialQueueAdapter::new(
                     trial_queue.clone_ref(py),
                 )));
                 Ok((queue, trial_queue.clone_ref(py)))
@@ -370,7 +370,7 @@ impl PyStudy {
         let directions: Vec<Direction> = directions.into_iter().map(|d| d.into()).collect();
         let (storage_arc, storage_pyobj) = Python::attach(|py| resolve_storage_pyobj(py, storage))?;
         let (sampler_arc, sampler_pyobj) = Python::attach(|py| resolve_sampler_pyobj(py, sampler))?;
-        let trial_queue = PyInMemoryTrialQueue ::new();
+        let trial_queue = PyInMemoryTrialQueue::new();
         let trial_queue_arc = trial_queue.queue.clone();
         let trial_queue_pyobj = Python::attach(|py| -> PyResult<Py<PyAny>> {
             Ok(Py::new(py, trial_queue)?.into_any())
