+++ rustuna_pyo3/src/trial_queue.rs
@@ -1,56 +1,18 @@
 use pyo3::prelude::*;
-use rustuna_core::trial_queue::TrialQueue;
-use rustuna_core::{Error, ErrorKind, Result};
 use std::sync::{Arc, RwLock};
+
+use crate::trial_queue::python::PythonTrialQueueAdapter;
 
 pub mod directory;
 pub mod inmemory;
+pub mod python;
 pub mod sqlite3;
-
-pub struct PyObjectTrialQueue {
-    obj: Py<PyAny>,
-}
-
-impl PyObjectTrialQueue {
-    pub fn new(obj: Py<PyAny>) -> Self {
-        Self { obj }
-    }
-
-    fn map_pyerr(&self, err: PyErr) -> Error {
-        Python::attach(|py| {
-            Error::with_reason(
-                ErrorKind::Unexpected,
-                format!("Python TrialQueueProtocol error: {}", err.value(py)),
-            )
-        })
-    }
-}
-
-impl TrialQueue for PyObjectTrialQueue {
-    fn enqueue(&mut self, trial_id: u32) -> Result<()> {
-        Python::attach(|py| {
-            self.obj
-                .call_method1(py, "enqueue", (trial_id,))
-                .map(|_| ())
-                .map_err(|err| self.map_pyerr(err))
-        })
-    }
-
-    fn dequeue(&mut self) -> Result<u32> {
-        Python::attach(|py| {
-            self.obj
-                .call_method0(py, "dequeue")
-                .and_then(|value| value.extract::<u32>(py))
-                .map_err(|err| self.map_pyerr(err))
-        })
-    }
-}
 
 #[derive(Clone)]
 #[pyclass(name = "PyObjectTrialQueue")]
 #[pyo3(module = "rustuna")]
 pub struct PyPyObjectTrialQueue {
-    pub queue: Arc<RwLock<PyObjectTrialQueue>>,
+    pub queue: Arc<RwLock<PythonTrialQueueAdapter>>,
 }
 
 #[pymethods]
@@ -58,7 +20,7 @@ impl PyPyObjectTrialQueue {
     #[new]
     fn new(trial_queue: Py<PyAny>) -> Self {
         Self {
-            queue: Arc::new(RwLock::new(PyObjectTrialQueue::new(trial_queue))),
+            queue: Arc::new(RwLock::new(PythonTrialQueueAdapter::new(trial_queue))),
         }
     }
 }
