+++ rustuna_pyo3/src/trial_queue/python.rs
@@ -1,0 +1,42 @@
+use pyo3::prelude::*;
+use rustuna_core::trial_queue::TrialQueue;
+use rustuna_core::{Error, ErrorKind, Result};
+
+pub struct PythonTrialQueueAdapter {
+    obj: Py<PyAny>,
+}
+
+impl PythonTrialQueueAdapter {
+    pub fn new(obj: Py<PyAny>) -> Self {
+        Self { obj }
+    }
+
+    fn map_pyerr(&self, err: PyErr) -> Error {
+        Python::attach(|py| {
+            Error::with_reason(
+                ErrorKind::Unexpected,
+                format!("Python TrialQueueProtocol error: {}", err.value(py)),
+            )
+        })
+    }
+}
+
+impl TrialQueue for PythonTrialQueueAdapter {
+    fn enqueue(&mut self, trial_id: u32) -> Result<()> {
+        Python::attach(|py| {
+            self.obj
+                .call_method1(py, "enqueue", (trial_id,))
+                .map(|_| ())
+                .map_err(|err| self.map_pyerr(err))
+        })
+    }
+
+    fn dequeue(&mut self) -> Result<u32> {
+        Python::attach(|py| {
+            self.obj
+                .call_method0(py, "dequeue")
+                .and_then(|value| value.extract::<u32>(py))
+                .map_err(|err| self.map_pyerr(err))
+        })
+    }
+}
